requirements.txt
delta-spark==1.0.0
pyspark==3.1.2
confluent-kafka==1.7.0